pyproject.toml
[project]
name = "hei-fastapi"
version = "1.0.0"
description = "Hei FastAPI Application - A Python implementation of hei-boot"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "CharlieZhang", email = "charlie@example.com"}
]
keywords = ["fastapi", "python", "hei-boot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.46.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.9.0",
    "sqlalchemy>=2.0.49",
    "pymysql>=1.1.1",
    "redis>=7.4.0",
    "pyyaml>=6.0.2",
    "python-multipart>=0.0.26",
    "pyjwt>=2.12.1",
    "bcrypt>=5.0.0",
    "gmssl>=3.2.2",
    "pillow>=12.1.1",
    "openpyxl>=3.1.5",
    "snowflake-id>=0.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "black>=25.0.0",
    "isort>=6.0.0",
    "flake8>=7.2.0",
    "mypy>=1.16.0",
]

[build-system]
requires = ["setuptools>=80.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "core*", "modules*"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
